requirements.txt
Django==1.7c2
Fabric==1.9.1
GDAL==1.11.1
Magic-file-extensions==0.2
Markdown==2.4.1
Pillow==2.6.0
PyYAML==3.11
defusedxml==0.4.1
django-activelink==0.4
django-annoying==0.7.9
django-bootstrap-form==3.1
django-debug-toolbar==1.0.1
django-filter==0.7
django-geojson==2.3.0
django-simple-captcha==0.4.1
django-tastypie==0.11.0
djangorestframework==2.3.14
docker-py==0.5.3
dockerpty==0.3.2
docopt==0.6.2
ecdsa==0.11
fig==1.0.1
futures==2.2.0
geopy==0.99
goslate==1.3.0
lxml==3.3.1
mercurial==3.1.1
nose==1.3.4
numpy==1.8.0
paramiko==1.14.0
pkginfo==1.2.1
psycopg2==2.5.2
pycrypto==2.6.1
pyenchant==1.6.5
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.4.3
six==1.8.0
sqlparse==0.1.11
suds==0.4
texttable==0.8.1
twine==1.5.0
vboxapi==1.0
virtualenv==1.11.6
websocket-client==0.11.0
wsgiref==0.1.2
wxPython==3.0.0.0
wxPython-common==3.0.0.0
